+++ src/main/java/br/com/infocedro/promocontrol/core/exception/UsuarioNaoEncontradoException.java
@@ -1,0 +1,8 @@
+package br.com.infocedro.promocontrol.core.exception;
+
+public class UsuarioNaoEncontradoException extends NotFoundBusinessException {
+
+    public UsuarioNaoEncontradoException() {
+        super("Usuario nao encontrado");
+    }
+}