+++ src/main/java/br/com/infocedro/promocontrol/infra/controller/dto/ResetarSenhaUsuarioRequest.java
@@ -1,0 +1,6 @@
+package br.com.infocedro.promocontrol.infra.controller.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record ResetarSenhaUsuarioRequest(@NotBlank String username) {
+}