+++ src/main/java/br/com/infocedro/promocontrol/infra/controller/dto/ResetarSenhaUsuarioResponse.java
@@ -1,0 +1,6 @@
+package br.com.infocedro.promocontrol.infra.controller.dto;
+
+public record ResetarSenhaUsuarioResponse(
+        String username,
+        String senhaTemporaria) {
+}